requirements.txt
asgiref==3.4.1
astroid==2.8.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coverage==6.2
cryptography==35.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.8
django-allauth==0.46.0
django-jenkins==0.110.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.3
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.1
platformdirs==2.4.0
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
pylint==2.11.1
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.13.2
